+++ docs/changelog.rst
@@ -10,7 +10,10 @@ Unreleased
 - Added complete support for HTML :py:class:`Tags <.Tag>`, along with
   appropriate unit tests. This includes forms like ``<ref>foo</ref>``,
   ``<ref name="bar"/>``, and wiki-markup tags like bold (``'''``), italics
-  (``''``), and lists (``''*''``, ``''#''``, ``'';''`` and ``'':''``).
+  (``''``), and lists (``*``, ``#``, ``;`` and ``:``).
+- :py:class:`Wikicode's <.Wikicode>` :py:meth:`.filter` methods are now passed
+  *recursive=True* by default instead of *False*. **This is a breaking change
+  if you rely on any filter() methods being non-recursive by default.**
 - Various fixes and cleanup.
 
 v0.2